organized programs pages and components & fixed design on view all programs page
