+++ frontend/src/models/programModel.ts
@@ -1,0 +1,10 @@
+export interface Program {
+	id: number;
+	name: string;
+	location: string;
+	description: string;
+	progress: number | null;
+	time: string | null;
+	date: string | null;
+	openSpots: number | null;
+}
