--- frontend/src/pages/Programs/ProgramComponents/TypeDefinitions/Program.ts
@@ -1,10 +1,0 @@
-export interface Program {
-	id: number;
-	name: string;
-	location: string;
-	description: string;
-	progress: number | null;
-	time: string | null;
-	date: string | null;
-	openSpots: number | null;
-}
