--- frontend/src/pages/Programs/ProgramPage/index.ts
@@ -1,1 +1,0 @@
-export { default as ProgramPage } from "./ProgramPage";
