--- frontend/src/pages/Programs/ProgramView/index.ts
@@ -1,1 +1,0 @@
-export { default as ProgramView } from "./ProgramView";
