+++ frontend/src/pages/Programs/components/AvailabilityGrid.tsx
