+++ frontend/src/pages/Programs/components/Description/Offer.tsx
