+++ frontend/src/pages/Programs/components/Description/Volunteering.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./Volunteering.css";
-import "../Description.css";
+import "../../Description/Description.css";
 
 function ProgramsVolunteering() {
 	return (
