+++ frontend/src/pages/Programs/components/DialogueBoxes/EducationDialogue.tsx
