+++ frontend/src/pages/Programs/components/DialogueBoxes/ReferencesDialogue.tsx
